Billing services unification

+++ codebase/commons/src/main/scala/com/virtuslab/payments/payments.scala
@@ -1,5 +1,5 @@
 package com.virtuslab.payments
 
 package object payments {
-  case class PaymentRequest(payer: String, payee: String, amount: Int)
+  case class PaymentRequest(payer: String, payee: String, amount: BigDecimal)
 }
